Moved Bookmarks directory to Media tab
